refactor and remove mock
