--- helix-gateway/src/main/java/org/apache/helix/gateway/mock/MockProtoResponse.java
@@ -1,15 +1,0 @@
-package org.apache.helix.gateway.mock;
-
-public class MockProtoResponse {
-
-  private String _messageId;
-
-  public MockProtoResponse(String messageId) {
-    System.out.println("Finished process of message : " + messageId);
-    _messageId = messageId;
-  }
-
-  public String getMessageId() {
-    return _messageId;
-  }
-}
